python/pyproject.toml
[tool.poetry]
name = "inclusive_code"
version = "0.1.0"
description = "A custom Pylint checker to enforce an inclusive language style guide"
authors = ["Helen Stockman <helen@flexport.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
pylint = "^2.8.2"
PyYAML = "^5.4.1"
pyhumps = "^1.6.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
